+++ app/public/app.css
@@ -1,3 +1,3 @@
-textarea.form-control {
+.form-group-sm textarea.form-control {
     height: 16em;
 }
